requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2==2.9.6
python-dotenv==1.0.0
pytz==2023.3
SQLAlchemy==2.0.16
typing_extensions==4.6.3
yarl==1.9.2
